// enums/PokemonTypes.ts
export enum PokemonTypes {
  Normal = 'Normal',
  Fighting = 'Fighting',
  Flying = 'Flying',
  Ground = 'Ground',
  Poison = 'Poison',
  Rock = 'Rock',
  Bug = 'Bug',
  Ghost = 'Ghost',
  Steel = 'Steel',
  Fire = 'Fire',
  Water = 'Water',
  Grass = 'Grass',
  Electric = 'Electric',
  Psychic = 'Psychic',
  Ice = 'Ice',
  Dragon = 'Dragon',
  Dark = 'Dark',
  Fairy = 'Fairy',
}